Precalculate frames for the edges.
When joining the meshes of each strut, their vertices on the cut-plane need
to line up. The best way to guarantee this is to make the cut-plane
projections of the radial axes of each strut/edge framee coincide.

I'm not sure the calcualtions are actually correct, but the script does run
without errors.