Redo find_edge_frame to use edges if faces aren't available.
If the edge has no suitable connected edges, then a suitable "up" vector is
selected, but with one good edge, then the normal of the plane defined by
the two edges is used. If two good edges are available, then the normals of
the two planes are bisected (same idea as using faces).